0.0.1 version of admin panel
